+++ optionals/no_stock_missions/$PBOPREFIX$
@@ -1,0 +1,1 @@
+x\cba\addons\no_stock_missions
