+++ optionals/no_stock_missions/no_stock_missions_prep/script_component.hpp
@@ -1,4 +1,3 @@
-#define COMPONENT nomissions
+#define COMPONENT no_stock_missions_prep
 #include "\x\cba\addons\main\script_mod.hpp"
-
 #include "\x\cba\addons\main\script_macros.hpp"
